+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `homebridge-nest-accfactory` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## v0.2.6 ()
+
+- Support for Nest Protect(s) in Google Home app
+- Logs Nest Protect(s) self testing status
+
 ## v0.2.5 (2024/12/10)
 
 - Fix for dropped sub modules.. Do not know why!